requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
python-decouple==3.8
requests==2.31.0
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pyyaml==6.0.1
swagger-ui-bundle==1.1.0 